requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
selenium==4.20.0
beautifulsoup4==4.12.3
jieba==0.42.1
